Update BotProfile.db
Adjusted the "Cost" of each "Skill Template" and now have more average skilled bots per team. So the chances of receiver a Highest Skill Bot and the Lowest Skilled bot are slim. It feels it balances out giving teams at least 1 Good and 1 Worst Bot with the others being averaged. This was tested on 5v5 Bot setup